requirements.txt
absl-py==2.1.0
aniso8601==9.0.1
attrs==23.2.0
bidict==0.22.1
blinker==1.7.0
cffi==1.16.0
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dnspython==2.5.0
eventlet==0.35.0
Flask==3.0.1
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
flatbuffers==23.5.26
fonttools==4.47.2
greenlet==3.0.3
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.4
matplotlib==3.8.2
mediapipe==0.10.9
numpy==1.26.3
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.0
pillow==10.2.0
protobuf==3.20.3
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
python-engineio==4.8.2
python-socketio==5.11.0
pytz==2023.3.post1
simple-websocket==1.0.0
six==1.16.0
sounddevice==0.4.6
tzdata==2023.4
Werkzeug==3.0.1
wsproto==1.2.0

python-dotenv~=1.0.1